requirements.txt
aioredis==1.3.1
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.5.7
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.7
cycler==0.10.0
daphne==3.0.2
Django==3.2
django-cleanup==5.2.0
fabio==0.11.0
h5py==3.2.1
hiredis==2.0.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
kiwisolver==1.3.1
matplotlib==3.4.2
msgpack==1.0.2
Pillow==8.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
requests==2.25.1
service-identity==21.1.0
silx==0.15.2
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
Twisted==21.2.0
txaio==21.2.1
typing-extensions==3.10.0.0
urllib3==1.26.4
websocket-client==1.0.1
zope.interface==5.4.0
